+++ src/app/components/section/section.component.ts
@@ -1,15 +1,17 @@
-import { SectionService } from "src/app/services/section.service";
-import { Component, OnInit } from "@angular/core";
-import { ISection } from "src/app/models/ISection";
-import { ActivatedRoute } from "@angular/router";
-import { PostTypes } from "src/app/constants/post-types.enum";
-import { IPost } from "src/app/models/IPost";
-import { PostService } from "src/app/services/post.service";
+import { Roles } from 'src/app/services/auth/roles.enum';
+import { AuthService } from 'src/app/services/auth/auth.service';
+import { SectionService } from 'src/app/services/section.service';
+import { Component, OnInit } from '@angular/core';
+import { ISection } from 'src/app/models/ISection';
+import { ActivatedRoute } from '@angular/router';
+import { PostTypes } from 'src/app/constants/post-types.enum';
+import { IPost } from 'src/app/models/IPost';
+import { PostService } from 'src/app/services/post.service';
 
 @Component({
-  selector: "app-section",
-  templateUrl: "./section.component.html",
-  styleUrls: ["./section.component.scss"],
+  selector: 'app-section',
+  templateUrl: './section.component.html',
+  styleUrls: ['./section.component.scss'],
 })
 export class SectionComponent implements OnInit {
   postTypes = PostTypes;
@@ -22,12 +24,13 @@ export class SectionComponent implements OnInit {
   constructor(
     private route: ActivatedRoute,
     private postService: PostService,
-    private sectionService: SectionService
+    private sectionService: SectionService,
+    private authService: AuthService
   ) {}
 
   ngOnInit() {
     this.route.paramMap.subscribe((params) => {
-      this.sectionId = params.get("id");
+      this.sectionId = params.get('id');
       this.loadSection();
       this.loadSectionPosts();
     });
@@ -48,19 +51,27 @@ export class SectionComponent implements OnInit {
       });
   }
 
-  addPost(post: IPost) {
+  addPost(postBody: IPost) {
     this.postService
-      .createPost(post, this.sectionId, this.creatingAtm)
+      .createPost(postBody, this.sectionId, this.creatingAtm)
       .subscribe((post) => {
         this.posts = [{ ...post, comments: [] }, ...this.posts];
       });
   }
 
   removePost(post: IPost) {
-    this.postService.deletePost(post._id).subscribe(() => {
-      this.posts = this.posts.filter(
-        (currentPosts) => currentPosts._id !== post._id
-      );
-    });
+    if (confirm('Are you sure you want to remove post?')) {
+      this.postService.deletePost(post._id).subscribe(() => {
+        this.posts = this.posts.filter(
+          (currentPosts) => currentPosts._id !== post._id
+        );
+      });
+    }
+  }
+
+  get canCreatePosts() {
+    return this.authService.currentUser.roles.some((role) =>
+      [Roles.ADMIN, Roles.PROFESSOR].includes(role as Roles)
+    );
   }
 }
